refactor(tui::ui::components::popups): move delete-confirm related popups to own file
