+++ tui/src/ui/components/popups/mod.rs
@@ -1,6 +1,5 @@
 #![allow(clippy::module_name_repetitions)]
 
-use crate::config::StyleColorSymbol;
 /**
  * MIT License
  *
@@ -24,22 +23,24 @@ use crate::config::StyleColorSymbol;
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-use crate::ui::{Id, Model, Msg};
+use crate::ui::Msg;
 use termusicplayback::SharedSettings;
-use tui_realm_stdlib::{Input, Paragraph, Radio};
-use tuirealm::command::{Cmd, CmdResult, Direction, Position};
-use tuirealm::event::{Key, KeyEvent, KeyModifiers, NoUserEvent};
+use tui_realm_stdlib::{Paragraph, Radio};
+use tuirealm::command::{Cmd, CmdResult, Direction};
+use tuirealm::event::{Key, KeyEvent, NoUserEvent};
 use tuirealm::props::{
-    Alignment, BorderType, Borders, Color, InputType, PropPayload, PropValue, TextModifiers,
-    TextSpan,
+    Alignment, BorderType, Borders, Color, PropPayload, PropValue, TextModifiers, TextSpan,
 };
 use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent, State, StateValue};
 
+mod deleteconfirm;
 mod error;
 mod help;
 mod podcast;
 mod saveplaylist;
 
+#[allow(unused_imports)]
+pub use deleteconfirm::{DeleteConfirmInputPopup, DeleteConfirmRadioPopup};
 pub use error::ErrorPopup;
 pub use help::HelpPopup;
 #[allow(unused_imports)]
@@ -167,185 +168,6 @@ impl Component<Msg, NoUserEvent> for QuitPopup {
 }
 
 #[derive(MockComponent)]
-pub struct DeleteConfirmRadioPopup {
-    component: Radio,
-    config: SharedSettings,
-}
-
-impl DeleteConfirmRadioPopup {
-    pub fn new(config: SharedSettings) -> Self {
-        let component = {
-            let config = config.read();
-            Radio::default()
-                .foreground(
-                    config
-                        .style_color_symbol
-                        .library_foreground()
-                        .unwrap_or(Color::LightRed),
-                )
-                // .background(Color::Black)
-                .background(
-                    config
-                        .style_color_symbol
-                        .library_background()
-                        .unwrap_or(Color::Reset),
-                )
-                .borders(
-                    Borders::default()
-                        .color(
-                            config
-                                .style_color_symbol
-                                .library_border()
-                                .unwrap_or(Color::LightRed),
-                        )
-                        .modifiers(BorderType::Rounded),
-                )
-                .title("Are sure you want to delete?", Alignment::Left)
-                .rewind(true)
-                .choices(&["No", "Yes"])
-                .value(0)
-        };
-
-        Self { component, config }
-    }
-}
-
-impl Component<Msg, NoUserEvent> for DeleteConfirmRadioPopup {
-    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
-        let config = self.config.clone();
-        let keys = &config.read().keys;
-        let cmd_result = match ev {
-            Event::Keyboard(KeyEvent {
-                code: Key::Left, ..
-            }) => self.perform(Cmd::Move(Direction::Left)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Right, ..
-            }) => self.perform(Cmd::Move(Direction::Right)),
-
-            Event::Keyboard(key) if key == keys.global_left.key_event() => {
-                self.perform(Cmd::Move(Direction::Left))
-            }
-            Event::Keyboard(key) if key == keys.global_right.key_event() => {
-                self.perform(Cmd::Move(Direction::Right))
-            }
-            Event::Keyboard(key) if key == keys.global_up.key_event() => {
-                self.perform(Cmd::Move(Direction::Left))
-            }
-            Event::Keyboard(key) if key == keys.global_down.key_event() => {
-                self.perform(Cmd::Move(Direction::Right))
-            }
-            Event::Keyboard(key) if key == keys.global_quit.key_event() => {
-                return Some(Msg::DeleteConfirmCloseCancel)
-            }
-            Event::Keyboard(key) if key == keys.global_esc.key_event() => {
-                return Some(Msg::DeleteConfirmCloseCancel)
-            }
-
-            Event::Keyboard(KeyEvent {
-                code: Key::Enter, ..
-            }) => self.perform(Cmd::Submit),
-            _ => return None,
-        };
-        if matches!(
-            cmd_result,
-            CmdResult::Submit(State::One(StateValue::Usize(0)))
-        ) {
-            Some(Msg::DeleteConfirmCloseCancel)
-        } else if matches!(
-            cmd_result,
-            CmdResult::Submit(State::One(StateValue::Usize(1)))
-        ) {
-            Some(Msg::DeleteConfirmCloseOk)
-        } else {
-            Some(Msg::None)
-        }
-    }
-}
-
-#[derive(MockComponent)]
-pub struct DeleteConfirmInputPopup {
-    component: Input,
-}
-
-impl DeleteConfirmInputPopup {
-    pub fn new(style_color_symbol: &StyleColorSymbol) -> Self {
-        Self {
-            component: Input::default()
-                .foreground(
-                    style_color_symbol
-                        .library_foreground()
-                        .unwrap_or(Color::Yellow),
-                )
-                .background(
-                    style_color_symbol
-                        .library_background()
-                        .unwrap_or(Color::Reset),
-                )
-                .borders(
-                    Borders::default()
-                        .color(style_color_symbol.library_border().unwrap_or(Color::Green))
-                        .modifiers(BorderType::Rounded),
-                )
-                // .invalid_style(Style::default().fg(Color::Red))
-                .input_type(InputType::Text)
-                .title("Type DELETE to confirm:", Alignment::Left),
-        }
-    }
-}
-
-impl Component<Msg, NoUserEvent> for DeleteConfirmInputPopup {
-    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
-        let cmd_result = match ev {
-            Event::Keyboard(KeyEvent {
-                code: Key::Left, ..
-            }) => self.perform(Cmd::Move(Direction::Left)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Right, ..
-            }) => self.perform(Cmd::Move(Direction::Right)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Home, ..
-            }) => self.perform(Cmd::GoTo(Position::Begin)),
-            Event::Keyboard(KeyEvent { code: Key::End, .. }) => {
-                self.perform(Cmd::GoTo(Position::End))
-            }
-            Event::Keyboard(KeyEvent {
-                code: Key::Delete, ..
-            }) => self.perform(Cmd::Cancel),
-            Event::Keyboard(KeyEvent {
-                code: Key::Backspace,
-                ..
-            }) => self.perform(Cmd::Delete),
-            Event::Keyboard(KeyEvent {
-                code: Key::Char(ch),
-                modifiers: KeyModifiers::SHIFT | KeyModifiers::NONE,
-            }) => self.perform(Cmd::Type(ch)),
-            Event::Keyboard(KeyEvent { code: Key::Esc, .. }) => {
-                return Some(Msg::DeleteConfirmCloseCancel);
-            }
-            Event::Keyboard(KeyEvent {
-                code: Key::Enter, ..
-            }) => self.perform(Cmd::Submit),
-            _ => CmdResult::None,
-        };
-        match cmd_result {
-            CmdResult::Submit(State::One(StateValue::String(input_string))) => {
-                if input_string == *"DELETE" {
-                    return Some(Msg::DeleteConfirmCloseOk);
-                }
-                Some(Msg::DeleteConfirmCloseCancel)
-            }
-            _ => Some(Msg::None),
-        }
-
-        // if cmd_result == CmdResult::Submit(State::One(StateValue::String("DELETE".to_string()))) {
-        //     Some(Msg::DeleteConfirmCloseOk)
-        // } else {
-        //     Some(Msg::DeleteConfirmCloseCancel)
-        // }
-    }
-}
-
-#[derive(MockComponent)]
 pub struct MessagePopup {
     component: Paragraph,
 }
@@ -375,33 +197,5 @@ impl Component<Msg, NoUserEvent> for MessagePopup {
         //     _ => None,
         // }
         None
-    }
-}
-
-impl Model {
-    pub fn mount_confirm_radio(&mut self) {
-        assert!(self
-            .app
-            .remount(
-                Id::DeleteConfirmRadioPopup,
-                Box::new(DeleteConfirmRadioPopup::new(self.config.clone())),
-                vec![]
-            )
-            .is_ok());
-        assert!(self.app.active(&Id::DeleteConfirmRadioPopup).is_ok());
-    }
-
-    pub fn mount_confirm_input(&mut self) {
-        assert!(self
-            .app
-            .remount(
-                Id::DeleteConfirmInputPopup,
-                Box::new(DeleteConfirmInputPopup::new(
-                    &self.config.read().style_color_symbol
-                )),
-                vec![]
-            )
-            .is_ok());
-        assert!(self.app.active(&Id::DeleteConfirmInputPopup).is_ok());
     }
 }
